+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/IRopeConnection.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.supplementaries.common.block;
 
+import net.mehvahdjukaar.amendments.mixins.LanternBlockPlacementMixin;
 import net.mehvahdjukaar.supplementaries.configs.CommonConfigs;
 import net.mehvahdjukaar.supplementaries.reg.ModTags;
 import net.minecraft.core.BlockPos;
@@ -8,6 +9,7 @@ import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.ChainBlock;
 import net.minecraft.world.level.block.FaceAttachedHorizontalDirectionalBlock;
+import net.minecraft.world.level.block.LanternBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.AttachFace;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
@@ -45,7 +47,7 @@ public interface IRopeConnection {
 
     static boolean canConnectDown(BlockState downState) {
         Block b = downState.getBlock();
-        if(b instanceof IRopeConnection ropeConnection){
+        if (b instanceof IRopeConnection ropeConnection) {
             return ropeConnection.canSideAcceptConnection(downState, Direction.UP);
         }
         return (downState.is(ModTags.ROPE_HANG_TAG)
@@ -55,7 +57,7 @@ public interface IRopeConnection {
     }
 
     static boolean isSupportingCeiling(BlockState upState, BlockPos pos, LevelReader world) {
-        if(upState.getBlock() instanceof  IRopeConnection ropeConnection){
+        if (upState.getBlock() instanceof IRopeConnection ropeConnection) {
             return ropeConnection.canSideAcceptConnection(upState, Direction.DOWN);
         }
         return (Block.canSupportCenter(world, pos, Direction.DOWN) && upState.isSolid()) || upState.is(ModTags.ROPE_SUPPORT_TAG);