+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/blocks/AbstractRopeBlock.java
@@ -6,7 +6,6 @@ import net.mehvahdjukaar.moonlight.api.misc.InvPlacer;
 import net.mehvahdjukaar.moonlight.api.util.Utils;
 import net.mehvahdjukaar.supplementaries.common.block.IRopeConnection;
 import net.mehvahdjukaar.supplementaries.common.block.ModBlockProperties;
-import net.mehvahdjukaar.supplementaries.common.utils.ItemsUtil;
 import net.mehvahdjukaar.supplementaries.common.utils.RopeHelper;
 import net.mehvahdjukaar.supplementaries.configs.CommonConfigs;
 import net.mehvahdjukaar.supplementaries.integration.CompatHandler;
@@ -19,7 +18,6 @@ import net.minecraft.sounds.SoundSource;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.ItemInteractionResult;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
@@ -36,7 +34,6 @@ import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
-import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.block.state.properties.BooleanProperty;
 import net.minecraft.world.level.material.Fluid;
 import net.minecraft.world.level.material.Fluids;
@@ -58,7 +55,7 @@ public abstract class AbstractRopeBlock extends WaterBlock implements IRopeConne
 
     public AbstractRopeBlock(Properties properties) {
         super(properties);
-                this.registerDefaultState(this.stateDefinition.any()
+        this.registerDefaultState(this.stateDefinition.any()
                 .setValue(KNOT, false).setValue(WATERLOGGED, false));
         shapes = this.makeShapes();
     }
@@ -91,7 +88,7 @@ public abstract class AbstractRopeBlock extends WaterBlock implements IRopeConne
 
     @Override
     public VoxelShape getCollisionShape(BlockState state, BlockGetter worldIn, BlockPos pos, CollisionContext context) {
-        if(!CommonConfigs.Functional.ROPE_HORIZONTAL.get())return Shapes.empty();
+        if (!CommonConfigs.Functional.ROPE_HORIZONTAL.get()) return Shapes.empty();
         return ((!hasConnection(Direction.UP, state) && (context.isAbove(COLLISION_SHAPE, pos, true) || !hasConnection(Direction.DOWN, state)))
                 || !(context instanceof EntityCollisionContext ec && ec.getEntity() instanceof LivingEntity) ?
                 getShape(state, worldIn, pos, context) : Shapes.empty());
@@ -259,7 +256,8 @@ public abstract class AbstractRopeBlock extends WaterBlock implements IRopeConne
                 }
             }
             if (!player.isShiftKeyDown() && hand == InteractionHand.MAIN_HAND) {
-                if (level.getBlockState(pos.below()).getBlock() == this) {
+                if (level.getBlockState(pos.below()).is(this)
+                        || level.getBlockState(pos.above()).is(this)) {
                     if (RopeHelper.removeRopeDown(pos.below(), level, this)) {
                         level.playSound(player, pos, SoundEvents.LEASH_KNOT_PLACE, SoundSource.BLOCKS, 1, 0.6f);
                         if (!player.getAbilities().instabuild) {